% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutType.R
\name{mutType}
\alias{mutType}
\title{Determine mutation type}
\usage{
mutType(VCFFile, refGenome, context_length, graphics = TRUE)
}
\arguments{
\item{VCFFile}{a set of mutations in VCF Format}

\item{refGenome}{Full reference genome sequences as provided by UCSC in
Biostrings objects}

\item{context_length}{an odd integer representing the length of region around
the mutation}

\item{graphics}{default = TRUE; if TRUE, saves a barplot of mutation type
frequencies in a pdf file}
}
\value{
vector of mutation types as "UP[REF>ALT]DOWN"
}
\description{
The mutType function compares a list of mutations in VCF format to a
reference genome and returns the mutation type along with upstream and
downstream bases for a total length = context_length
}
\examples{

library(BSgenome.Hsapiens.UCSC.hg19)
Hs <- Hsapiens
sample <- system.file("extdata", "sample.vcf", package = "mutType")

mutType(sample, Hs, 7)

}
\author{
Alessandro Giulivo\cr Politecnico di Milano\cr Maintainer: Alessandro
Giulivo\cr E-Mail: \href{mailto:alessandro.giulivo@mail.polimi.it}{alessandro.giulivo@mail.polimi.it}
}
