% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutTypeTable.R
\name{mutTypeTable}
\alias{mutTypeTable}
\title{Summarize Mutation Types}
\usage{
mutTypeTable(mutTypeResult)
}
\arguments{
\item{mutTypeResult}{a data.frame as obtained from the mutType function}
}
\value{
a table along with a histogram summarizing mutation types
}
\description{
The mutTypeTable function summarizes the mutations obtained from mutType
function into a counts table and produces a pdf image plotting the results.
It is automatically run when calling mutType function when
parameter graphics = T
}
\examples{

library(BSgenome.Hsapiens.UCSC.hg19)
Hs <- Hsapiens
sample <- system.file("extdata", "sample.vcf", package = "mutType")

muts <- mutType(sample, Hs, 7, graphics = FALSE)
mutTypeTable(muts)

}
\references{
\url{https://github.com/AlessandroGiulivo/sciprog/tree/main/RProj}
}
\seealso{
\code{\link{mutType}}
}
\author{
Alessandro Giulivo\cr Politecnico di Milano\cr Maintainer:
Alessandro Giulivo\cr E-Mail: \href{mailto:alessandro.giulivo@mail.polimi.it}{alessandro.giulivo@mail.polimi.it}
}
